refactor: [WIP] place cell functions into hook
